requirements.txt
Package               Version
--------------------- ----------------
acme                  1.21.0
apturl                0.5.2
attrs                 21.2.0
bcrypt                3.2.0
blinker               1.4
Brlapi                0.8.3
certbot               1.21.0
certifi               2020.6.20
chardet               4.0.0
click                 8.0.3
colorama              0.4.4
command-not-found     0.3
ConfigArgParse        1.5.3
configobj             5.0.6
cryptography          3.4.8
cupshelpers           1.0
dbus-python           1.2.18
defer                 1.0.6
distro                1.7.0
distro-info           1.1+ubuntu0.2
docker                5.0.3
docker-compose        1.29.2
dockerpty             0.4.1
docopt                0.6.2
duplicity             0.8.21
environs              9.5.0
fasteners             0.14.1
future                0.18.2
grpcio                1.53.0
gyp                   0.1
httplib2              0.20.2
idna                  3.3
importlib-metadata    4.6.4
jeepney               0.7.1
josepy                1.10.0
jsonschema            3.2.0
keyring               23.5.0
language-selector     0.1
launchpadlib          1.10.16
lazr.restfulclient    0.14.4
lazr.uri              1.0.6
lockfile              0.12.2
louis                 3.20.0
macaroonbakery        1.3.1
Mako                  1.1.3
MarkupSafe            2.0.1
marshmallow           3.23.1
milvus-lite           2.4.10
monotonic             1.6
more-itertools        8.10.0
netifaces             0.11.0
numpy                 2.1.3
oauthlib              3.2.0
olefile               0.46
packaging             24.2
pandas                2.2.3
paramiko              2.9.3
parsedatetime         2.6
pexpect               4.8.0
Pillow                9.0.1
pip                   22.0.2
pip-review            1.3.0
protobuf              5.29.0
ptyprocess            0.7.0
pycairo               1.20.1
pycups                2.0.1
PyGObject             3.42.1
PyICU                 2.8.1
PyJWT                 2.3.0
pymacaroons           0.13.0
pymilvus              2.2.8
PyNaCl                1.5.0
pyOpenSSL             21.0.0
pyparsing             2.4.7
pyRFC3339             1.1
pyrsistent            0.18.1
python-apt            2.4.0+ubuntu4
python-dateutil       2.9.0.post0
python-debian         0.1.43+ubuntu1.1
python-dotenv         1.0.1
pytz                  2022.1
pyxdg                 0.27
PyYAML                5.4.1
reportlab             3.6.8
requests              2.25.1
requests-toolbelt     0.9.1
SecretStorage         3.3.1
setuptools            75.6.0
six                   1.16.0
systemd-python        234
texttable             1.6.4
tqdm                  4.67.1
tzdata                2024.2
ubuntu-drivers-common 0.0.0
ubuntu-pro-client     8001
ufw                   0.36.1
ujson                 5.10.0
unattended-upgrades   0.1
urllib3               1.26.5
usb-creator           0.3.7
wadllib               1.3.6
websocket-client      1.2.3
wheel                 0.37.1
xdg                   5
xkit                  0.0.0
zipp                  1.0.0
zope.component        4.3.0
zope.event            4.4
zope.hookable         5.1.0
zope.interface        5.4.0
